Search half works!!
rn if you type say, ACCOUNTING, you'll get any course with that contained in the name. The one slight caveat is that it only applies to the max of ten courses on the current page you are on, not the entire collection of courses. Also, rn if you type COMP, you'll get any course with COMP in the title, but it DOESN'T currently apply to course code as well (example, iOS programming would NOT show up, even though it is a COMP class). But this can be changed in the near future.