More search functionality
You can enter a term like ACCOUNTING in search bar as well as ACCT 201 A (or subsections of that full code) and get what you would expect. Note, for the filter to work correctly, you have to type in the search bar after entering filters. Perhaps this should be tweaked sligthly.